Create MatchingColors.js
This code uses the spreadsheet provided as a reference. If you have one table you'd like to apply this code to, you can just add the range (var range) and sheet (var sh) directly into the code. However, if you're using several instances of this code, it's best to make a separate function that invokes this as many times as you require.