Create Rich-Text-Cells-2.js
This particular code does what the read-me describes, but the way you decide formatting is with word order/number. This is good for if a word will be ever-changing, but not if you're going to be adding a any text before the selected word. So, it's good for if you have a cell reference in the original cell and you just want that cell reference's displayed data formatted. Keep in mind, it's order, not index, so the first number is 1!