Create CopyDisplayValue.js
Example sheet: https://docs.google.com/spreadsheets/d/1qHPG1M36KFVbRNdeYoiU5wxQAmoQvq9JmBE331eFWbw/edit?usp=sharing

The code is initiated with defaults, but please feel free to change the original values.

This code is also made so that you can call it as many times as needed, for different cells. Copied display value(s) is/are able to be formatted as much as wanted.